life/pyproject.toml
[project]
name = "life"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.11.3",
    "pyyaml>=6.0.2",
    "rich>=14.0.0",
]
